forgot to commit this
git-svn-id: https://svn.code.sf.net/p/qterm/code/trunk/qterm-qt4@1178 bd9efd21-7b0e-0410-bb59-ddbc1b2ac886

+++ src/qtermcanvas.h
@@ -39,9 +39,9 @@ protected slots:
     void ccwRotate();
     void deleteImage();
 protected:
-    void moveImage(float, float);
-    void resizeImage(float);
-    void rotateImage(float);
+    void moveImage(double, double);
+    void resizeImage(double);
+    void rotateImage(double);
 
     void closeEvent(QCloseEvent *);
     void mousePressEvent(QMouseEvent*);
